requirements.txt
fastapi==0.70.0
uvicorn==0.11.5
alembic==1.7.4
SQLAlchemy==1.3.13
PyYAML==5.3.1
SQLAlchemy-Utils==0.36.8
psycopg2-binary==2.8.5
jinja2
python-socketio==4.6.0
aiohttp==3.8.0
requests
python-socketio[asyncio_client]
python-engineio==3.14.2
aio-pika==6.8.0